+++ .phpstorm.meta.php/phpunit.meta.php
@@ -13,25 +13,25 @@ namespace PHPSTORM_META {
 
     override(\PHPUnit\Framework\TestCase::createConfiguredMock(0),
         map([
-            '' => '@&\PHPUnit\Framework\MockObject\MockObject',
+            '' => '@',
         ])
     );
 
     override(\PHPUnit\Framework\TestCase::createMock(0),
         map([
-            '' => '@&\PHPUnit\Framework\MockObject\MockObject',
+            '' => '@',
         ])
     );
 
     override(\PHPUnit\Framework\TestCase::createPartialMock(0),
         map([
-            '' => '@&\PHPUnit\Framework\MockObject\MockObject',
+            '' => '@',
         ])
     );
 
     override(\PHPUnit\Framework\TestCase::getMockForAbstractClass(0),
         map([
-            '' => '@&\PHPUnit\Framework\MockObject\MockObject',
+            '' => '@',
         ])
     );